backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1

# 数据库驱动
aiosqlite==0.19.0
pymysql==1.1.0
aiomysql==0.2.0
cryptography==41.0.7

# 配置管理
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP请求
httpx==0.25.2
requests==2.31.0

# 企业微信
wechatpy==1.8.18
xmltodict==0.13.0

# AI服务
anthropic==0.7.8  # Claude API
google-generativeai==0.3.2  # Gemini API

# 图像处理 (OCR功能)
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10

# 任务调度
celery==5.3.4
redis==5.0.1

# 数据处理
pandas==2.1.3
python-dateutil==2.8.2

# 工具库
python-dotenv==1.0.0
validators==0.22.0
typing-extensions==4.8.0

# 日志
loguru==0.7.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1

# 定时任务调度
apscheduler==3.10.4

# 系统监控
psutil==5.9.6